Made the volatility graph reusable. Renamed and moved almost all files related to the graph. Added coloring options.
