+++ src/components/portfolio/DetailsMainAnalytics.tsx
@@ -1,12 +1,14 @@
 import React from 'react';
 import { Theme } from '@material-ui/core';
-import { createStyles, makeStyles } from '@material-ui/core/styles';
+import { createStyles, makeStyles, useTheme } from '@material-ui/core/styles';
 import { useTranslation } from 'react-i18next';
 import { NonEmptyPortfolioDetails } from '../../portfolio/APIClient';
 import StyledNumberFormat from '../shared/StyledNumberFormat';
 import DetailsAnalyticsHeatmap from './DetailsMainAnalyticsHeatmap';
 import DetailsAnalyticsDebtEquityBar from './DetailsAnalyticsDebtEquityBar';
-import DetailsVolatilityGraph from './DetailsVolatilityGraph';
+import DetailsVolatilityGraph from '../shared/VolatilityGraph';
+import VolatilityGraphPersonalPortfolio from '../shared/LargeVolatilityLineEntry';
+import VolatilityLineEntry from './VolatilityLineEntry';
 
 const useStyles = makeStyles(({ palette }: Theme) =>
   createStyles({
@@ -26,11 +28,30 @@ type DetailsMainAnalyticsProps = {
   portfolio: NonEmptyPortfolioDetails;
 };
 
+type VolatilityLineData = {
+  [symbol: string]: number;
+};
+
 const DetailsMainAnalytics: React.FC<DetailsMainAnalyticsProps> = ({
   portfolio,
 }) => {
+  const { palette } = useTheme();
   const classes = useStyles();
   const { t } = useTranslation();
+
+  const sortedStocks: VolatilityLineData = {};
+
+  if (portfolio.positions) {
+    Object.values(portfolio.positions)
+      // sort by total value (largest to smallest)
+      .sort((a, b) => b.qty * b.stock.price - a.qty * a.stock.price)
+      // take the top 5
+      .slice(0, 4)
+      .map((p) => p.stock)
+      .forEach((s) => {
+        sortedStocks[s.symbol] = s.volatility;
+      });
+  }
 
   return (
     <>
@@ -63,10 +84,23 @@ const DetailsMainAnalytics: React.FC<DetailsMainAnalyticsProps> = ({
       <div className={classes.titleWrapper}>
         {t('portfolio.details.analytics.volatility.vsMarket')}
       </div>
-      <DetailsVolatilityGraph
-        positions={portfolio.positions}
-        portfolioVolatility={portfolio.analytics.volatility}
-      />
+      <DetailsVolatilityGraph color={palette.primary.contrastText}>
+        <VolatilityGraphPersonalPortfolio
+          volatilityValue={portfolio.analytics.volatility}
+          name={t(
+            'portfolio.details.analytics.volatility.myPortfolio'
+          ).toString()}
+          textColor={palette.primary.contrastText}
+        />
+        {Object.entries(sortedStocks).map(([symbol, volatility]) => (
+          <VolatilityLineEntry
+            key={symbol}
+            volatilityValue={volatility}
+            tooltipText={symbol}
+            color={volatility > 1 ? palette.error.main : palette.success.main}
+          />
+        ))}
+      </DetailsVolatilityGraph>
     </>
   );
 };
